+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [2.1.0]
 
+### Features
+
+- One keyframe for Fov Layer is now parsed and executed
+
 ### Fixes
 
 - Can't create characters in NeoForge dedicated server 1.21.1